requirements.txt
google-genai
openai
pydantic
google-auth
gspread
requests
python-dotenv
pdf2image
Pillow
schedule
